completely restructured backend to account for tests, some test functionality working
