+++ Backend/index.js
@@ -1,83 +1,32 @@
 const express = require('express')
 const mongoose = require('mongoose')
-const Schema = mongoose.Schema
-var fs = require('fs')
-const { response } = require('express')
-const cors = require('cors')
-const bodyParser = require('body-parser')
-const { nextTick } = require('process')
 
+const bodyParser = require('body-parser')
+const cors = require('cors')
 
 const app = express()
 app.use(cors({origin:true}))
-
 app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: true }))
 
-
-const movieSchema = new Schema 
-(
-    {
-        'id':Number,
-        'title': String,
-        'release date': String,
-        'starring': String,
-        'genre': String,
-        'runtime': Number,
-        'rating': String,
-        'thumbnail': String
-    }
-)
-
-const locationSchema = new Schema 
-(
-    {
-        'id': Number,
-        'address': Number,
-        'city': String,
-        'opening times': Number,
-        'telephone': String,
-        'sessions': [Object]
-    }
-)
-
-const bookingSchema = new Schema 
-(
-    {
-        'location': Number,
-        'date': String,
-        'screen': Number,
-        'seat': String,
-        'reference': String
-    }
-)
-
-const priceSchema = new Schema 
-(
-    {
-        'type': String,
-        'price': Number
-    }
-)
-
-// Testing replies collection
-const repliesSchema = new Schema
-(
-    {
-        '_id': Number,
-        'content': String,
-        'content': [Array]
-    }
-)
-
-const Movie = mongoose.model('Movie', movieSchema)
-const Location = mongoose.model('Location', locationSchema)
-const Booking = mongoose.model('Booking', bookingSchema)
-const Price = mongoose.model('Price', priceSchema)
-const Replies = mongoose.model('Replies', repliesSchema)
+app.use(express.json())
+app.use(express.static('script'));
 
 const uri = 'mongodb://localhost:27017/qacinemas';
 let opts = { useNewUrlParser: true }
+
+// Routes
+const bookingRouter = require("./routes/bookingRoutes");
+const locationRouter = require("./routes/locationRoutes");
+const movieRouter = require("./routes/movieRoutes");
+const priceRouter = require("./routes/priceModels");
+const repliesRouter = require("./routes/repliesModels");
+
+app.use("/", bookingRouter);
+app.use("/", locationRouter);
+app.use("/", movieRouter);
+app.use("/", priceRouter);
+app.use("/", repliesRouter);
 
 mongoose.connect
 (uri, opts,
@@ -92,469 +41,10 @@ mongoose.connect
             console.log("connection ready")
         }
     }
-)
+);
 
-app.use(express.json())
-app.use(express.static('script'));
+const myserver = app.listen(4000, () =>
+    console.log("Server started on", myserver.address().port)
+);
 
-app.get
-('/movielist', function(request, response, error)
-    {
-        Movie.find
-        ({}, (error, movies) =>
-            {
-                if (error)
-                {
-                    console.log('Error movielist: ' + error)
-                    response.send([{"id":0, "title":'Error movielist: ' + error}])
-                }
-                else
-                {
-                    if (movies == '')
-                    {
-                        response.send([{"id":0, "title":"Error movielist: there is no movie collection, please create one"}])
-                    }
-                    else
-                    {
-                        let list = []
-
-                        movies.forEach
-                        ((movie) =>
-                            {
-                                list.push(movie)
-                            }
-                        )
-
-                        response.send(list)
-                    }
-                }
-
-
-            }
-        ).sort('title')
-    }
-)
-
-app.get
-('/nowshowing', function(request, response)
-    {
-        Movie.find
-        ({'release date':''}, (error, movies) =>
-            {
-                let list = []
-
-                if (error)
-                {
-                    console.log('Error movielist: ' + error)
-                    list.push({"id":0, "title":'Error movielist: ' + error})
-                }
-                else
-                {
-                    if (movies == '')
-                    {
-                        list.push({"id":0, "title":'Error movielist: there is no movie collection, please create one'})
-                    }
-                    else
-                    {
-                        movies.forEach
-                        ((movie) =>
-                            {
-                                list.push(movie)
-                            }
-                        )
-                    }
-                }
-
-                response.send(list)
-            }
-        ).sort('title')
-    }
-)
-
-app.get
-('/comingsoon', function(request, response)
-    {
-        Movie.find
-        ({"release date":{ "$ne": ""}}, (error, movies) =>
-            {
-                let list = []
-
-                if (error)
-                {
-                    console.log('Error movielist: ' + error)
-                    list.push({"id":0, "title":'Error comingsoon: ' + error})
-                }
-                else
-                {
-                    if (movies == '')
-                    {
-                        list.push({"id":0, "title":'Error comingsoon: There is no movie collection, please create one'})
-                    }
-                    else
-                    {
-                        movies.forEach
-                        ((movie) =>
-                            {
-                                list.push(movie)
-                            }
-                        )
-                    }
-                }
-
-                response.send(list)
-            }
-        ).sort('title')
-    }
-)
-
-
-
-app.get
-('/movie/select/:id', function(request, response)
-    {
-        let id = request.params.id
-
-        Movie.find
-        ({'id':id}, (error, movie) =>
-            {
-                if (error)
-                {
-                    console.log('Error /movie/select/' + id + ': ' + error)
-                    movie = [{"id":0, "title":'Error /movie/select/ ' + id + ': ' + error}]
-                }
-
-                if (movie == '')
-                {
-                    movie = [{"id":0, "title":'Cannot find a movie title with the id: ' + id}]
-                }
-
-                response.send(movie)
-            }
-        )
-    }
-)
-
-app.get
-('/movie/search/genre/:genre', function(request, response)
-    {
-        let genre = request.params.genre
-
-        Movie.find
-        ({'genre':genre}, (error, movies) =>
-            {
-                let list = []
-
-                if (error)
-                {
-                    console.log('Error /movie/select/genre/'+ genre + ': ' + error)
-                    list.push({"id":0, "title":'Error /movie/select/genre/' + genre + ': ' + error})
-                }
-                else
-                {
-                    if (movies == '')
-                    {
-                        list.push("No movie found for genre: " + genre)
-                    }
-                    else
-                    {
-                        movies.forEach
-                        ((movie) =>
-                            {
-                                list.push(movie)
-                            }
-                        )
-                    }
-                }
-                
-                response.send(list)
-            }
-
-        )
-    }
-)
-
-app.get
-('/movie/search/title/:title', function(request, response)
-    {
-        let title = request.params.title
-
-        Movie.find
-        ({'title':{$regex: title, $options: 'i'}}, (error, movies) =>
-            {
-                let list = []
-
-                if (error)
-                {
-                    console.log('Error /movie/select/title/'+ title + ': ' + error)
-                    list.push({"id":0, "title":'Error movie/select/title/' + title + ': ' + error})
-                }
-                else
-                {
-                    if (movies == '')
-                    {
-                        list.push("No movie title found using the search criteria: " + title)
-                    }
-                    else
-                    {
-                        movies.forEach
-                        ((movie) =>
-                            {
-                                list.push(movie)
-                            }
-                        )
-                    }
-                }
-                
-                response.send(list)
-            }
-
-        )
-    }
-)
-
-app.get
-('/locationlist', function(request, response)
-    {
-        Location.find
-        ({}, (error, locations) =>
-            {
-                let list = []
-
-                if (error)
-                {
-                    console.log('Error locataionlist: ' + error)
-                    list.push({"id":0, "city":'Error locationlist: ' + error})
-                }
-                else
-                {
-                    if (locations == '')
-                    {
-                        list.push({"id":0, "city":'Error locationlist: there is no location collection, please create one'})
-                    }
-                    else
-                    {
-                        locations.forEach
-                        ((location) =>
-                            {    
-                                list.push(location)
-                            }
-                        )
-                    }
-                }
-
-                response.send(list)
-            }
-        ).sort('city')
-    }
-)
-
-app.get
-('/location/select/:id', function(request, response)
-    {
-        let id = request.params.id
-        let movies = []
-
-        Movie.find
-        ({}, (error, movielist) =>
-            {
-                if (error)
-                {
-                    console.log('Error location/select/' + id + ' Movie.find: ' + error)
-                }
-                else
-                {
-                    movielist.forEach
-                    ((movie) =>
-                        {    
-                            movies[movie.id] = movie.title
-                        }
-                    )
-                }
-            }
-        )
-
-        Location.find
-        ({"id":id}, (error, locations) =>
-            {
-                let location = ''
-
-                if (error)
-                {
-                    console.log('Error location/select/' + id + ': ' + error)
-                    location = {"id":"Error location/select: " + id + " : " + error}
-                }
-                else
-                {
-                    let loc = locations[0]
-                    let sessions = []
-                    
-                    for (let i in loc.sessions)
-                    {
-                        let session = loc.sessions[i]
-                        let id = loc.sessions[i].movie
-                        
-                        if (movies[id])
-                        {
-                            session.movie = {"id":id, "title":movies[id]}
-                        }
-                        else
-                        {
-                            session.movie = {"id":id, "title":'Movie id: ' + id + ' cannot be found on the movie collection' }
-                        }
-
-                        sessions.push(session)
-                    }
-
-                    location = {"id": loc.id, "city": loc.city, "sessons":sessions}
-                }
-
-                response.send(location)
-            }
-        )
-    }
-)
-
-
-
-app.get
-('/bookings/', function(request, response)
-    {
-        Booking.find
-        ({}, (error, bookinglist) =>
-            {
-                let bookings = []
-
-                if (error)
-                {
-                    console.log('Error bookings:' + error)
-                    bookings.push({"location":"Error bookings: " + error})
-                }
-                else
-                {
-                    for (let i in bookinglist)
-                    {
-                        bookings.push(bookinglist[i])
-                    }
-                }
-
-                response.send(bookings)
-            }
-        )
-    }
-)
-
-app.get
-('/bookings/select/location/:id', function(request, response)
-    {
-        let id = request.params.id
-
-        Booking.find
-        ({"location":id}, (error, bookinglist) =>
-            {
-                let bookings = []
-
-                if (error)
-                {
-                    console.log('Error bookings/select/location/' + id + ' ' + error)
-                    bookings.push({"location":"Error bookings/select/location/" + id + " + error"})
-                }
-                else
-                {
-                    bookings.push(bookinglist)
-                }
-
-                response.send(bookings)
-            }
-        )
-    }
-)
-
-app.get
-('/prices', function(request, response)
-    {
-        Price.find
-        ({}, (error, pricelist) =>
-            {
-                let prices = []
-
-                if (error)
-                {
-                    console.log('Error prices ' + error)
-                    prices.push({"error":"Error prices" + error})
-                }
-                else
-                {
-                    for (let i in pricelist)
-                    {
-                        prices.push(pricelist[i])
-                    }
-                }
-
-                response.send(prices)
-            }
-        )
-    }
-)
-
-app.post
-('booking/add', function(request, response)
-    {
-        let booking = request.body.content
-
-        Console.log('booking ' + booking)
-
-        Booking.insertOne(booking, (error) =>
-            {
-                if (error)
-                {
-                    Console.log('Error booking/add ' + error)
-                    response.send('Error booking/add ' + error)
-                }
-                else
-                {
-                    response.send(0)
-                }
-            }
-        )
-    }
-)
-
-
-// Global variable for id
-let id = 1;
-
-// Reply builder for replies
-let replyBuilder = (replyContent, replyID) => {
-    let reply = {
-        _id : replyID,
-        content : replyContent
-    }
-    return reply;
-}
-
-
-// Post reply function
-app.post("/replies/createReply", (req, res, next) => {
-    let newReply = replyBuilder(req.body.content, parseInt(id));
-    id++;
-    // let newReply = req.body;
-    Replies.create(newReply)
-      .then((result) => res.status(201).send(result))
-      .catch((err) => next(err));
-});
-
-// View all replies
-app.get("/replies/readReplies", (req, res, next) => {
-    Replies.find()
-      .then((results) => res.send(results))
-      .catch((err) => next(err));
-  });
-
-// Delete a reply
-app.delete("/replies/delete/:id", (req, res, next) => {
-    Replies.findByIdAndDelete(req.params.id)
-    .then(() => res.status(200).send())
-    .catch((err) => next(err));
-  });
-
-app.listen(4000);
+module.exports = myserver;
