+++ Backend/models/locationModels.js
@@ -1,0 +1,17 @@
+const mongoose = require('mongoose')
+const Schema = mongoose.Schema
+
+const locationSchema = new Schema 
+(
+    {
+        'id': Number,
+        'address': Number,
+        'city': String,
+        'opening times': Number,
+        'telephone': String,
+        'sessions': [Object]
+    }
+);
+
+const Location = mongoose.model('Location', locationSchema);
+module.exports = Location;
